% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACTIONetExperiment-package.R, R/AllClasses.R
\docType{package}
\name{ACTIONetExperiment}
\alias{ACTIONetExperiment}
\alias{ACTIONetExperiment-class}
\title{ACTIONet: Robust multi-resolution analysis of single-cell datasets}
\usage{
ACTIONetExperiment(
  ...,
  rowNets = S4Vectors::SimpleList(),
  colNets = S4Vectors::SimpleList(),
  rowMaps = S4Vectors::SimpleList(),
  colMaps = S4Vectors::SimpleList()
)
}
\arguments{
\item{...}{SummarizedExperiment and SummarizedExperiment components}

\item{rowNets, colNets}{gene-gene and cell-cell networks, respectively}

\item{rowMaps, colMaps}{Factorization results (W and H matrices)}
}
\value{
an ACTIONetExperiment (ACE) object

An ACTIONetExperiment (ACE) object, derived from SummarizedExperiment, with additional slots to store ACTIONet results
}
\description{
Joint archetypal/network-based analysis for multi-resolution decomposition of single-cell states.
}
\section{Slots}{

\describe{
\item{\code{rowNets,colNets}}{gene-gene and cell-cell networks, respectively}

\item{\code{rowMaps,colMaps}}{Factorization results (W and H matrices)}
}}

\section{Usage}{


\enumerate{
\item ?reduce.sce to reduce raw count matrix in an SummarizedExperiment objects.
\item ?run.ACTIONet to run ACTIONet on SummarizedExperiment objects.
}
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/shmohammadi86/ACTIONet/issues}
\item Read the manuscript:
\href{https://www.biorxiv.org/content/10.1101/746339v2}{A multiresolution framework to characterize single-cell state landscapes}.
}
}

\author{
Shahin Mohammadi
}
