% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get.R, R/set.R
\name{colNets}
\alias{colNets}
\alias{colNets,ACTIONetExperiment-method}
\alias{colNets<-,ACTIONetExperiment-method}
\title{Get colNets}
\usage{
colNets(object, ...)

\S4method{colNets}{ACTIONetExperiment}(object)

\S4method{colNets}{ACTIONetExperiment}(object) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of adjacency matrices

List of adjacency matrices
}
\description{
Get colNets

Get column-associated networks

Set column-associated networks
}
