% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ace.R
\name{filter.ace.by.attr}
\alias{filter.ace.by.attr}
\title{Filter columns and rows of `ACTIONetExperiment` or `SummarizedExperiment` object by column attribute.}
\usage{
filter.ace.by.attr(
  ace,
  by,
  assay_name = "counts",
  min_cells_per_feat = 0.003,
  min_feats_per_cell = 1000,
  min_umis_per_cell = NULL,
  max_umis_per_cell = 50000,
  max_mito_fraction = 10,
  species = "hsapiens"
)
}
\description{
Filter columns and rows of `ACTIONetExperiment` or `SummarizedExperiment` object by column attribute.
}
