% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get.R, R/set.R
\name{rowMaps}
\alias{rowMaps}
\alias{rowMaps,ACTIONetExperiment-method}
\alias{rowMaps<-,ACTIONetExperiment-method}
\title{Get rowMaps}
\usage{
rowMaps(object, ...)

\S4method{rowMaps}{ACTIONetExperiment}(object, all = TRUE)

\S4method{rowMaps}{ACTIONetExperiment}(object) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of matrices

List of matrices
}
\description{
Get rowMaps

Get row-associated factors

Set row-associated factors
}
