% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get.R, R/set.R
\name{rowMapTypes}
\alias{rowMapTypes}
\alias{rowMapTypes,ACTIONetExperiment-method}
\alias{rowMapMeta,ACTIONetExperiment-method}
\alias{rowMapTypes<-,ACTIONetExperiment-method}
\title{Get rowMapTypes}
\usage{
rowMapTypes(object, ...)

\S4method{rowMapTypes}{ACTIONetExperiment}(object, all = TRUE)

\S4method{rowMapMeta}{ACTIONetExperiment}(object, all = TRUE)

\S4method{rowMapTypes}{ACTIONetExperiment}(object) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of types

List of types

List of matrices
}
\description{
Get rowMapTypes

Get row-associated factor types

Get row-associated factor metadata

Set row-associated factor types
}
