% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get.R, R/set.R
\name{rowNets}
\alias{rowNets}
\alias{rowNets,ACTIONetExperiment-method}
\alias{rowNets<-,ACTIONetExperiment-method}
\title{Get rowNets}
\usage{
rowNets(object, ...)

\S4method{rowNets}{ACTIONetExperiment}(object)

\S4method{rowNets}{ACTIONetExperiment}(object) <- value
}
\arguments{
\item{object}{ACTIONetExperiment object}

\item{...}{other parameters}
}
\value{
List of adjacency matrices

List of adjacency matrices
}
\description{
Get rowNets

Get row-associated networks

Set row-associated networks
}
